Add _execfile function
This critical function is the heart of the integration.

The code reads the contents of the tempfile passed over the socket,
compiles the contents into a code_obj via the 'compile' python builtin,
redirects stdout into a stringbuffer, execs the code in the
interpreters's context, resets stdout, echos the result server-side, and
then returns the output.

Special mega thanks to Ian Thompson's deep python knowledge for this
monster:

    exec code_obj in __main__.__dict__

which was the way to force the exec to happen in the same context as
the nuke script editor.